Arduino.h: addMillisHrs now takes a float

+++ include/arduino-mock/Arduino.h
@@ -87,12 +87,12 @@ public:
   void addMillisRaw (unsigned long milliseconds) {currentMillis += milliseconds;};
   void setMillisSecs(unsigned long seconds) {setMillisRaw(seconds*      1000);};
   void setMillisMins(unsigned long minutes) {setMillisRaw(minutes*   60*1000);};
-  void setMillisHrs (unsigned long hours)   {setMillisRaw(hours  *60*60*1000);};
+  void setMillisHrs (float         hours)   {setMillisRaw(hours  *60*60*1000);};
 
   void setMillisRaw (unsigned long milliseconds) {currentMillis = milliseconds;};
   void addMillisSecs(unsigned long seconds) {addMillisRaw(seconds*      1000);};
   void addMillisMins(unsigned long minutes) {addMillisRaw(minutes*   60*1000);};
-  void addMillisHrs (unsigned long hours)   {addMillisRaw(hours  *60*60*1000);};
+  void addMillisHrs (float         hours)   {addMillisRaw(hours  *60*60*1000);};
 
   
   MOCK_METHOD2(pinMode, void (uint8_t, uint8_t));
